+++ app/src/main/java/com/app/body_manage/data/calendar/Weekday.kt
@@ -8,6 +8,6 @@ data class Weekday(
     override val hasLunch: Boolean = false,
     override val hasDinner: Boolean = false,
     override val hasMiddle: Boolean = false,
-    override val kcal: Int,
+    override val kcal: Long,
     override val weight: Float?,
 ) : Day
