+++ app/src/main/java/com/app/body_manage/ui/top/TopViewModel.kt
@@ -100,7 +100,7 @@ class TopViewModel(
         }
     }
 
-    fun setGoalKcal(goal: Int) {
+    fun setGoalKcal(goal: Long) {
         viewModelScope.launch {
             userPreferenceRepository.setGoatKcal(goal)
             load()
